remove middleware
